+++ desmume/src/cheatSystem.h
@@ -36,7 +36,7 @@ typedef struct
 	u32		hi[255];
 	u32		lo[255];
 	char	description[75];
-	u8		num;
+	int		num;
 	u8		size;
 
 } CHEATS_LIST;
